requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
jinja2==3.1.2
python-multipart==0.0.6
pdfplumber==0.6.0
python-dotenv==1.0.0
httpx==0.27.0
